Parser Class: execute user input and return the response
The execute function includes all the details of action for all kind of Commands. It causes the method to be extremely long.

Let's create a Command class and make all different kind of commands to be subclasses of Command class, and then abstract details of action into the runCommand function of each subclasses.

Using Command class reduce the length of Parser.execute function and improve the readability. SLAP is achieved as well.