Add Java Assertions
Currently, there are not assertions for the methods in the package. This
means that the integrity of the code is not checked as the code runs,
and there could be unintended bugs at various points in the code. This
might result in regression bugs in the future when the code is extended.

To test the codebase, Java assertions were added at various points to
check that the methods are working as intended.

Lets
* Add assertions to ensure that the code is working at various points