class TaskList: Improve readability of listOut(), exportOut(), listFeature(), findFeature() methods
There are complicated expression in the conditional statements of these methods that can be written better to improve readability.

Removing these complicated expressions and rephrasing them would make the method easier to read.

As a step towards better readability, let's extract the complicated expressions and store them as variables, which can then be used in the conditional statements. Doing so will improve readability of the methods.