Fix bugs for handling events and finding tasks
Current implementation does not prompt user if user forgets to input dates for events, and does not tell user that no tasks are found if there's no relevant results.

This does not give a clear course of action for the user.

Let's
Include exception handling for the case where events have missing dates.
Include a message for the user if there are no relevant tasks to be found.