class TaskList: Update importIn() method, Create doWithinPeriodFeature() method
In order to incorporate the DoWithinPeriod task, we have to create a doWithinPeriodFeature that handles the user input for that case and creates a DoWithinPeriod object and adds it into the list.

The importIn() method needs to be updated to incorporate for the format that the DoWithinPeriod object is written into data/yale.txt file. We can then extract the data from yale.txt containing the DoWithinPeriod task and add it to the list such that its data is saved every time we close the program.

Doing so would allow for the DoWithinPeriod object to be saved and loaded automatically as the program is started.