Add assertions
Index of task to get is not checked.

Invalid index results in program crashing.

Checking the index before passing it to get() allows us to pass an
error message to the user instead of crashing the pogram.

Let's assert the index to make sure it is valid before passing it to
get().

Using assert allows us to document the assumption that index is valid.