Improve code quality
Some methods are too long and some are no longer used. Proper coding
standards are also not being adhered to in some parts.

These methods may hinder code readability making it harder for other
people to understand certain parts of the code.

Let's,
* remove methods/ classes no longer in use
* refactor long methods
* adhere to coding standard

Removing Ui package and Ui.java as it is no longer in use after
implementing GUI.

Refactoring methods in Storage as they are too long and hard to read.

Renaming done to isDone in Task.java as it is more readable.