Refactor Duke to decouple chatbot logic and UI
Duke logic is tightly coupled to UI and makes assumptions about the
implementation details of the UI class (e.g. that it behaves like a
terminal, or that it can block and wait for user input).

This makes implementing other types of user interfaces (e.g. desktop
GUIs or web UIs) more cumbersome than necessary, such as displaying
different types of messages in different styles.

Representing the input/output to/from the chatbot logic as abstract
messages allows us to accomodate a larger variety of behaviors and
quirks for different UIs, and possibly easily expand the types of
messages that can be displayed in different styles in the future.

Refactoring chatbot logic out into a discrete class also allows us to
add support for different types of UIs with relative ease, and will
facilitate easier testing (as we can test the chatbot logic separately
from the UI logic).