Fix similarity cutoff for task find
The command "find" used to search for task uses a fuzzy string matching
algorithm using Levenshtein distances between strings to calculate
similarities.

The chosen strategy of matching is a partial match, as to provide a
substring search as well. The output of the algorithm is between 0 and
100, hence a fix to the cutoff was required.

75 is chosen to be the threshold as it was experimented to be the most
forgiving while at the same time requires a "close enough" match.