Fix bug in parsing strings representing deadlines and events in local storage file
Assertion errors are thrown from the factory methods in Deadline and Event classes indicating that the strings representing deadlines and events in the local storage file do not have the correct date and time format. There is a mistake in the assert statement - it assumes that the strings in the local storage follow the format for user input when in fact the user input format and the file format are different. After fixing this mistake, another bug is discovered - the parsing of deadlines and events from the local storage file is incorrect and inadvertently adds an extra white space between the date and time.

To prevent the assertion error, the incorrect assert statements should first be removed. Then, to ensure that the deadlines and events are parsed correctly from the storage file, the relevant methods in the Deadline and Event classes should be fixed.

Let's remove the assert statements in the factory methods of Deadline and Event. Let's also change the methods used to parse the date and time of deadlines and events in the Deadline and Event classes.

The assert statements in the factory methods of Deadline and Event are temporarily removed until a good method to check if the strings in the local storage file follow the file format is found.