Add BetterGUI
GUI does not visually differentiate error message and bot responses. GUI images is a little too big which result in lesser space for the responses.

Add error message checking which allows the highlight of error messages, turning dialog box to a light pink color. Modify images to be rounded and reduced in size.

This enables user to easily differentiate error messages due to the color differences and allows more information to be displayed at once due to the smaller image.

Other info
- Add window title
- Rename method names (proper naming)