TaskList class: change the find method to enable better search (partial matching of keywords)
The logic for executing the Find command is inside the find method of the TaskList class. The current find method only returns tasks that match all the inputted keywords. This limits the function of find as users may enter multiple keywords and the number of tasks that contain all those keywords in the inputted order may be few / none.

Changing the find method in the TaskList class to return tasks that match one or more of the keywords instead of all of them increases the number of matching tasks returned.

Let's change the find method in the TaskList class to add matching tasks for each inputted keyword to the returned string.

Returning the tasks that match one or more keywords improves the search quality and efficiency (users do not have to enter multiple Find commands for each keyword).