Improve GUI of Doge Chatbot
Tweak the GUI to match the asymmetric nature of the conversation. As the conversation is between the user and the app (not between two humans), it is asymmetric in nature. Accordingly, it makes sense not to display both sides of the conversion in the same visual format.

Changing the background colours of the user and chatbot chat bubbles and also flipping the dialog box to match the asymmetric nature of the conversation.

By changing the colours and flipping the dialog boxes, it allows the user to clearly identify and differentiate the respective chat bubbles.