Fix Minor Bugs
Current implementation has bugs such as not throwing an error when a task does not exist in the list and invalid commands such as "todolist"

This confuses the user and would bring about bad user experience.

Let's
Throw an exception when the task does not exist.
Include a space in the command to be queried in the startsWith method.