Edit execute() method in Parser class to follow coding standard.
The method execute() does not follow the coding standard due to its
length. There are common logic in the method which cause
code duplication.

Extracting this common logic allows the method to be significantly
shorter and the principle of SLAP to be followed. For example,
the method parseByCondition() and parseString() shows the
SLAP principle.