Users can add tasks freely, without being notified if they have any other engagements that day.
If there are any clashes the user can be notified so that they are not
double-booked for different tasks at the same time.

Add a feature to warn the users about possible clashes, by modifying the
AddCommand class. New methods have to be added to TaskList, Ui, and
Tasks and its children classes.

Modifying AddCommand class lets us deal with the issue at the
appropriate level.