User could add similar tasks.
This means that users could keep adding the same tasks, and this means
that conan stores duplicates of the same task.

Let's Modify the code so that new tasks added are not duplicates. If
they are duplicates then issue a warning to the user, and ask them to
delete the previous task if they would like to add the new task.

Modifying AddCommand and Ui class would help us deal with the problem at
the appropriate level.