Improve the GUI
Currently, the user interface only has text in a vertical box.
To make it more aesthetically pleasing, we need to adjust and
match the asymmetric nature of the conversation.

We add a profile picture and a dialog box flip to emulate
user and virtual assistant interaction.

This helps users differentiate between a command and a response.