Update Task Types
Storage contains the types (declared as constant Strings)
of the task for importing declared independently from the Task classes.

The common value of the types across the Task classes and Storage class
cause code duplication.

The constants are declare in the sub-classes of the Task classes and
the getType() method is made public so that the storage class can access
the values.

This is to ensure that the types used in Storage during import is
consistent with the types used during export.