Add assert statements to important assumptions.
Currently, the program lacks any Assert statements. However, it is
important for programmers not to assume the code will always work
as intended.

This commit adds assert statements to document important assumptions.
Java assert statements have been used so that the failed assertions will
be detected when the program is run with Gradle.

Note: be sure to enable assertions in the build.gradle file to verify
assertions