Add Assertions
Add Assertions for index out of bounds exception for delete/mark/unmark commands.

Had to handle this exception for user error, change made addresses this.

It is done using assertions to ensure index is within bounds.

Change some of the code behind the bye command and storage.

GUI was not exiting properly when bye was called, and some of the writing to storage had issues.

Add a new method rewriteData() to address this issue and fixed bye command.

Change storage writing for marked vs unmarked to 1(marked) vs 0(unmarked), since the GUI had issues displaying the tick.