Add assert statements to Cleese and Ui classes
The Cleese and Ui class are crucial to get the program to a
state where we can test basic functionality and correctness.

Adding assert statements will help to ensure that the classes
and underlying functions are always given the correct input.

This allows for increased confidence of bugs or issues
arising from other parts of the program and speeds up
potential debugging.