Add OOP to code: add a UI manager class
The Main class is in charge of initializing the GUI when the app is launched; there is no clear separation between the logic and ui components of the app.

Adding a Ui class and transferring the code from Main's start method to this new class and making AnnBot the "Main" class in charge of the app's logic ensures that there is a separation between the logic and ui components of the app.

Let's add a Ui class which manages the GUI components of the app and convert AnnBot into the Main class by letting it extend Application and implement the start method for launching the app.

Increasing the code's OOP makes it easier to maintain the code, as changes to the implementation of the UI will not affect the logic of the app.