Changes to the controllers of the GUI components: allow users to exit the app using the command line interface (CLI)
Currently, users cannot exit from the app using the CLI as there is no method to handle a "bye" command in the MainWindow class. The "bye" command is handled in a similar way to other commands - a DialogBox containing AnnBot's response to the command is displayed on the screen.

Adding an method to handle the "bye" command in the MainWindow class will allow users to exit the app using the CLI.

Let's change the MainWindow class to handle the "bye" command differently from other commands.