Add assertions in some various points in code
Assertions are required in order to check whether the code
is able to work.

Assertions allow assertion error to be thrown if some conditions
are not held properly. For example, implementing assertion task
cannot be null in the method showAddedMessage(Task task, String
noOfTask) will throw assertion error if the input of task is null.