Update command response
Current command responses is not intuitive enough that the user will be able to figure out what the proper format of the commands are.

User might not be able to use the program as intended due to the lack of information in the response message.

Updating the messages to be clearer by displaying what information is missing/erroneous in the command. Also added examples to make it clearer.