User is confused on the commands to use for the application
User will not be able to use the application as intended and deter its
user experience

HelpCommand class is added to ensure that the user is able to see the
list of commands to be use in the application. Adjust welcome message to
ensure that the help command option is seen by the user

This ensures that the user's first touch point with the application
would be able to see that there is a help command as an option to use.
This would make the user experience better and use the application as
intended by the development team.