Delete dead codes.
There are several dead codes in the code base since a GUI had
been implemented. Most of them are located in Duke.java and Ui.java.

Since a GUI had been implemented, the way to display Duke's responses
has changed. Instead of using System.out.println(), the string of
response needs to be returned to be displayed on the GUI.