Add User Guide and representative screenshot
Current documentation is insufficient for user to get enough information to use the application effectively.

As such user may be unable to use the application as smoothly as intended.

Adding a user guide and a representative screenshot will give the user something to refer to to get a brief overview and guide of the program.