Change Parser class to add more OOP and remove magic values
There is a lot of hard code and code repetition in the Parser class which makes it difficult to change the format of the commands in the future.

Removing the magic values (numbers and Strings) and creating constants in the Command classes and the Parser class to store these values make it easier to change these values in the future.

Let's extract the arbitrary values in the code of the Parser class and assign them to constants in the Parser class and the Command classes.