Task not saved properly and no dynamic file location
User's task would not be saved if they close the application abruptly,
only way was using command bye. File location to be saved is not
dynamic, user might not have data/ folder in their system. This will
cause the retriving and saving of files function to break.

All commands that require to be saved to text file will write to file
after every action. Dynamic folder and file location has been added to
prevent issues with retriving and saving of files.

This is to ensure that the functionality does not break no matter which
system the application is hosted on.