Update storage path
Currently the storage path defined might be ambiguous and cause bugs in certain OSes without properly defined JVM path.

Can cause data to be saved in unintended locations and as a result lose the tasks that has been saved.

Properly defined the saving directory to be in the same directory as the jar file.