Improve code quality : deep nesting in Task
Class Task has a code quality issue with regards to deep nesting.

Using if else statement in this case will confuse reader. There is also an issue of deep nesting within the if else statements which is a code quality violation.

Let's refactor the code and use switch case instead to represent the conditional statements. The implementation of switch case and if statements does not need much change after all.

Switch case will be more intuitive to other developers as it clearly states what the essence of the conditions.