Update Storage file saving
Current file saving does not work as intended with directories that contains spaces.

If user saves the jar file in a directory containing spaces, the data file may be saved incorrectly to another directory.

Updating the storage file path to take into account spaces, this should make saving data work as intended regardless of space or not.