Change mainClassName in build.gradle
Currently, the mainClassName is indicated as duke.Duke

This needs to be changed as the JavaFX feature is now added to the
program. After adding JavaFX to Duke, the new Main class of the
code is the Launcher class.

Changed the mainClassName indicated to duke.Launcher. This is to
generate a new fat JAR file with JavaFX included in the JAR file.