Add Assertions to program
The code did not have any assertions. The assumptions about the
program state should be checked and verified at runtime.

Additional checking helps to ensure that code does not have any
errors.

Add to specific methods where there is a clear assumption about
the value of a certain variable.

There are clear assumptions here that should be checked.