Update Code Quality
Some parts of the code did not follow the java coding standard.
Some parts of the code had methods that exceed 60LoC.
Event and deadline tasks did not have a specific check if "/at"
and "/by" were present respectively in the input.

To follow the coding standard. To make the code more readable and
easily more understandable.
This needs to change so that users have a better understand of what
is missing from their inputs.

Break up long codes into smaller chunk.
Add Exception check in both parseEvent and parseDeadline methods.

Using an exception ensures that the error can be found and picked
out.