Add Help command to Duke
Users need to know all the commands of the program in order to use
the application. New Users will be at a lose.

First time users of Duke need a way to seek help whenever they are
stuck. Even experienced users may need a reminder on how to use
Duke.

Help command is added. String of all commands in Duke and what they
are used for.

Just a manual for Users to refer to whenever needed.