Add Assertions
Event and Deadline tasks does not check for assumptions that
string parsed as date and time are of correct size.

Event and Deadline tasks would cause Jeff to crash if certain
assumptions are not met.

Let's add assertions on those very variables have been added,
hence throwing an assertion failure instead of adding an
invalid entry into the list.

Assertions impact on performance is considered low compared
to its benefits.