Add a new extension to jBot.
User can now type in "today" to retrieve tasks they have today.

A simple keyword "today" is used for it to be easily remembered and
quickly typed in when user are in a rush.

Date on the saved deadlines and events are compared with today's date and
then listed out as a list.