JUnit testing: Created new JUnit tests for a more comprehensive test
Several refactoring were made previously, and tests and stubs
are unable to perform a comprehensive test on the system.

Several JUnit tests were created, specifically for the smaller
methods. These tests are important as it is used widely across
the code base.

It is done this way to improve the comprehensiveness of the testing
unit.

Let's,
* think of better ways to increase the scope of the other JUnit
tests.