Add assertions to classes in command, duke and task packages
There is no current use of assertions throughout the methods.

There are no complementary checks on the program state during runtime
apart from the use of exceptions, which can miss out on potential bugs.

Let's add assertions to the methods with parameter arguments needed to
check for null arguments and invalid integer values which should not be
passed in.

These assertions will help to verify the program state during run time
to ensure that the code is behaving as it is assumed to be.