Add method in TaskList class to search for any event clashes
There are no checks for date and time clashes when a user tries to
add an event task to the task list.

Multiple events held at the same date will be allowed to be added
to the existing task list, which does not showcase good 'task
management' functionality.

Let's add a findClashes method in the TaskList class to return true
if the date argument passed in matches any of the dates of existing
event tasks in the TaskList object.

Within this commit, bugs and errors in the Parser class are fixed
as well to fix the functionality of the task manager.