Modify code for parsing of input command and output display
The format of user input for adding event and deadline classes uses
" - at:" and " - by:" respectively. This format is standard across
the Command package class methods as well as in TaskList class.

We want to modify the output display to such that the dates are
placed within brackets. We want to modify the user input format
to accept " /at " and " /by ". This variation is subjective, to
improve the user display and input formatting.

Let's modify the following code:
* Add a toWrite() method in the Task class and subclasses to ensure
  the text to be written to the taskHistory.txt file is standardised
  (i.e. using " - at :" and " - by:")
* Modify the way user command is split in the command subclasses for
  AddEventCommand and AddDeadlineCommand to abide by the new variation
  (i.e. using " /at " and "/by "
* Modify the toString() method in the Task class and subclasses to
  ensure a standard format with dates placed within brackets for the
  Event and Deadline class