Add Parser.java to handle the logic of the program when interacting with the user
There are several different valid commands the program can take as
inputs, in some cases the commands are not inputted in a valid syntax.

Consolidating the majority of the program logic into Parser.java instead
of lumping them with the main method improves code quality and follows
coding standards.

As a step towards improving code quality and utilising OOP, let's
abstract the code that handles the program logic into Parser.java. Doing
so will make code readability and quality better.