Add assertions to avoid invalid input
There are several methods which require the input to be valid. This
includes requiring a Task to have a description, a Deadline to have
a description and a due date, an Event to have a description and a
venue and a valid task ID for update and get operations.

To do so, let’s include assert statements to ensure that these
conditions are met before we carry out the methods.