The current program does not make use of the Assert features.
Assert function to be added to the program to test certain hypothesis in the event of failure.

The number of characters for a list command should be fixed. The assert feature will verify the same and inform the user of any errors.

Lets,
Add the assert feature to the the Duke bot to ensure we are able to run the program smoothly.