Duke class:  skeleton framework of the program
Methods are factored out into super classes.

The common behaviors across the program is to parse the input from the user.

Extracting the common behavior into a super class allows us to use
polymorphism when dealing with multiple inputs.
For example same method can deal with various inputs.

Let's use SLAP to identify the method to be factored out, which is Parser in this instance