Detect duplicate tasks
Users may forget that they have already added a task to their list and
thus may create duplicate tasks in the list. To avoid this, let’s add a
method to check if two Task objects are equal and another method
to check if the list already contains this task.