use assert feature to document assumptions
Not using assertions at various code checkpoints like the creation of
the file and empty input commands to ensure that assumptions hold.

Adding assertions allows to properly manage the assumptions and provide
a feedback system if the code stops working.

Important code fragments that depend on external system are being
asserted of the presence of a particular assumption in place before
performing any operations.

Assert checks whether the assumptions made by the developer are correct
or not.