perform critical code checks and improve the code quality
The common behaviors across deadline and event input handlers cause code duplication.

Extracting the common behaviors into a method allows us to abstract out
the granularities of checking the input command.

Let's pull up behaviors common to the method which checks and
initializes important parameters for the tasks.

Making a method is preferrable here because there are only two types of
tasks that make use of this parameter check to avoid invalid commands.