Add catch block for DateTimeParseException in command subclasses
AddDeadlineCommand and AddEventCommand classes do not have a catch
block to deal with dates that have been input in the wrong format.

User will not be able to view the error message on the GUI window
if the dates entered for the command has not been parsed in properly.

Let's add a catch block in the two classes above, to deal with the
DateTimeParseException. The catch block will throw a new
DukeException, and and display a message to tell the user what is
the right date format to input.
a