Wrap words in message bubbles
Some messages can be long enough that the resulting message bubble
exceeds the width of the window.

An obvious fix would have been to enable word wrapping. However, the
old MessageBubble implementation did not have the ability to resize
dynamically, as the sizes were determined on creation (based on the
bounds of the text).

We reimplement MessageBubble to allow them to resize dynamically,
which also allows us to wrap (and... unwrap?) words easily.