Modify command classes to handle exception for index out of bounds
TaskList class currently uses assertions to check that the task id
passed in is a valid index of the existing task list.

There is no message displayed back to the user in the case where the
task id is invalid.

Let's modify the method body in the TaskList class to remove the
assertions. Lets add a catch block in the command subclasses to handle
index array out of bounds exception to throw a new Duke exception which
will be handled.