build.gradle : Update dependancies in the file to add JavaFX dependencies
To implement JavaFX and incorporate gui in our code, we need to add the related dependencies for it.

We add the corresponding dependencies to our build.gradle so we are able to use JavaFX and incorporate gui in our code.

We need to add dependencies to our build.gradle so we can use gradle to build our gui, this is the most efficient way.