Improve code quality
There are many areas of code quality to ensure that one's program is
considered a fine piece of work.

Amongst the various dimensions of code quality, such as run-time
efficiency, security, and robustness, one of the most important is
understandability.

Thus let us make an effort to maintain code quality not just for others
to understand, but also for ourselves as we might become 'strangers'
to our own code someday.