Add override annotation for overridden method in Command child classes
Child classes of Command class currently do not have override
annotation for run() method.

Compiler is not able to check for overridden methods. This could
potentially lead to overlooked errors when implementing the methods in
child classes which are expected to override the method in the parent
class.

Let's add the overridden annotation to relevant methods in the child
classes.