Attempt to fix occasional thread blocking on bye command
Calls to Thread.sleep() causes blocking effects on JavaFX
application thread. It appears that adding the Juke response to
the bye command is not thread safe, and occasionally the
application exits without showing the bye command due to the
thread block. Attempted to use JavaFX concurrency, however, the
problem still persists albeit at a lower frequency.