rectify bugs in snooze
local variable newString in method snooze of class TaskList is
initialised to an empty string.

works well when criteria for snooze is "both" date and time. However, in
the case of "time" only, the newTime variable does not get appropriately
updated leading to parsing errors.

intialising newTime directly to information[2] so that if the change
criteria is "time" only then an empty string will not be parsed.
If change criteria is "both" then newTime will be appropriately
updated to data[1].

Since the steps of updating date and time overlap with updating time
only, it is more logical to, conditionally, update date first. In such
an event, the date is separated from the input and then passed into
newTime for time updation. In the event that date is not meant to be
updated, the newTime string remains empty and leads to errors parsing.
Therefore, if it is directly initialised to the needed argument for
setTime() then errors can be evaded.