SavedDataParser: Add a new method isMarkedTask()
Rationale:
There was a bug where task marked as done was not saved properly in the text file (thanks Charisma). This was due to the parse() only adding unmarked tasks back to the tasklist despite their marked status. parse() is then modified accordingly to include isMarkedTask() checks. Marked tasks are added accordingly.