Implement Stats command
Duke needs a way to keep track of tasks completed in a given time frame.

A Stats command is useful as it allows users to keep track of their progress over time,
not just as a one-off event.

Let's update the Parser and TaskList to identify and process Stats command. Additionally, lets add a createdDate variable
to all the tasks.

The createDate variable will allow us to keep track of a task's creation relative to the current date, thus displaying completion
relative to current day.